+++ src/Loggers/Logger.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rubix\Server\Loggers;
+
+use Psr\Log\AbstractLogger;
+
+abstract class Logger extends AbstractLogger
+{
+    //
+}